fix(spi_nand_flash): correct copyright headers on OOB layout files
Replace mikkeldamsgaard project 2022 headers with Espressif Systems 2026 as

the primary copyright holder, matching reviewer feedback (PR#7 C4).

+++ spi_nand_flash/priv_include/nand_oob_layout_types.h
@@ -57,7 +57,7 @@ typedef struct {
  * (see default layout implementation notes).
  */
 typedef struct spi_nand_oob_layout {
-    uint8_t oob_bytes;
+    uint16_t oob_bytes;
     spi_nand_oob_bbm_t bbm;
     const spi_nand_ooblayout_ops_t *ops;
 } spi_nand_oob_layout_t;
